+++ lib/Db/V11/IndexManager.php
@@ -6,12 +6,12 @@ declare(strict_types=1);
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
-namespace OCA\Polls\Db\V10;
+namespace OCA\Polls\Db\V11;
 
 use Doctrine\DBAL\Schema\Exception\IndexAlreadyExists;
 use Doctrine\DBAL\Schema\Exception\IndexDoesNotExist;
 use Exception;
-use OCA\Polls\Migration\V10\TableSchema;
+use OCA\Polls\Migration\V11\TableSchema;
 use OCP\IConfig;
 use OCP\IDBConnection;
 use Psr\Log\LoggerInterface;