+++ lib/Migration/Version080301Date20250822153002.php
@@ -8,8 +8,8 @@ declare(strict_types=1);
 
 namespace OCA\Polls\Migration;
 
-use OCA\Polls\Db\V10\IndexManager;
-use OCA\Polls\Db\V10\TableManager;
+use OCA\Polls\Db\V11\IndexManager;
+use OCA\Polls\Db\V11\TableManager;
 use OCP\DB\ISchemaWrapper;
 use OCP\IDBConnection;
 use OCP\Migration\IOutput;