+++ CHANGELOG.md
@@ -11,6 +11,11 @@ All notable changes to this project will be documented in this file.
 ### Added
  - OpenAPI spec
 
+### Fixed
+ - Fixed some code inconsitencies regarding nullish groupIds and pollIds.
+ - Added share repair fix to repair steps after updates.
+ - Purge soft-deleted entries after the documented 12 hour undo window
+
 ### Changes
  - Added salt to user name obfuscation
  - Support Nextcloud 31